+++ Engine/Poo.h
@@ -6,19 +6,19 @@
 class Poo
 {
 public:
-    void Init(int in_x, int in_y, int in_vx, int in_vy);
+    void Init(float in_x, float in_y, float in_vx, float in_vy);
     void Update();
     void ProcessConsumption(const Dude& dude);
     void Draw(Graphics& gfx) const;
     bool IsEaten() const;
 
 private:
-    int x;
-    int y ;
-    int vx;
-    int vy;
-    static constexpr int width = 24;
-    static constexpr int height = 24;
+    float x;
+    float y ;
+    float vx;
+    float vy;
+    static constexpr float width = 24.0f;
+    static constexpr float height = 24.0f;
     bool isEaten = false;
     bool initialized = false;
 };
